requirements.txt
Django==2.0.4
django-polymorphic==2.0.2
psycopg2-binary==2.7.4
pytz==2018.4
